funread_backend/requirements.txt
aenum==3.1.11
Django==4.0.2
djangorestframework==3.13.1
djangorestframework_simplejwt==5.3.0
drf_extra_fields==3.4.1
drf-yasg==1.21.7
PyJWT==2.8.0
django-cors-headers==3.11.0
Pillow==10.0.1
whitenoise==6.0.0
mysqlclient==2.1.1
pygame==2.5.1
gTTS==2.3.2
googletrans==4.0.0rc1
python-dotenv==1.0.0
torch==2.0.1
tiktoken==0.3.3
moviepy==1.0.3
librosa==0.10.1
git+https://github.com/openai/whisper.git #instalar whisper

#Si ocurre algun problema con whisper usar este dos comandos
#pip install --upgrade --no-deps --force-reinstall git+https://github.com/openai/whisper.git